requirements.txt
python-dotenv
langchain-openai
langchain_community
langchain_core
streamlit
openai